% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\docType{data}
\name{input}
\alias{input}
\title{Dane wejsciowe dla funkcji \code{emisjadk()}}
\format{
Ramka danych zawiera:
\describe{
\item{Nat}{Natezenie ruchu}
\item{Segment}{Rodzaj nadwozia samochodu}
\item{Fuel}{Rodzaj uzywanego paliwa}
\item{Technology}{Technologia pracy silnika}
}
}
\usage{
input
}
\description{
Dane wejsciowe dla funkcji \code{emisjadk()}
}
\examples{
\dontrun{
input
}
}
\keyword{datasets}
