% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wskazniki.R
\docType{data}
\name{wskazniki}
\alias{wskazniki}
\title{Ramka danych emisji spalin}
\format{
Ramka danych zawiera:
\describe{
\item{Category}{Kategoria pojazdu - character}
\item{Fuel}{Rodzaj paliwa - character}
\item{Segment}{Rodzaj podwozia - character}
\item{Euro.Standard}{Norma dopuszczalna emisji spalin - character}
\item{Technology}{Technologia pracy silnika - character}
\item{Pollutant}{Rodzaj zanieczyszczenia - character}
\item{Mode}{Tryb jazdy pojazdu - character}
\item{Alpha}{parametr Alpha - double}
\item{Beta}{parametr Beta - double}
\item{Gamma}{parametr Gamma  - double}
\item{Delta}{parametr Delta - double}
\item{Epsilon}{parametr Epsilon - double}
\item{Zita}{parametr Zita - double}
\item{Hta}{parametr Hta - double}
\item{Reduction}{parametr Reduction - double}
\item{Bio}{parametr Bio - double}
\item{Procent}{parametr Procent - double}
}
}
\usage{
wskazniki
}
\description{
Dane zawierajace informacje o emisjach spalin pojazdow dla roznych kategorii.
}
\examples{
\dontrun{
 wskazniki
 }
}
\keyword{datasets}
