% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykresdk.R
\name{wykresdk}
\alias{wykresdk}
\title{Funkcja,ktora tworzy wykres na podstawie obliczonych emisji spalania}
\usage{
wykresdk(dane = out, filtr = "Pollutant")
}
\arguments{
\item{dane}{dataframe - dane wynikowe z funkcji emisjadk()}

\item{filtr}{character - filtr kolumny kategorii}
}
\value{

}
\description{
Funkcja tworzy wykres na podstawie danych wynikowych z funkcji \code{emisjadk()}.
Danymi wejsciowymi jest ramka danych \code{out} oraz parametr \code{filtr} ,dzieki ktorym
wybieramy kolumne kategorii np. Pollutant, Technology, Euro Standard.
}
\examples{
1. wykresdk(out, "Pollutant")
2. wykresdk(out, "Technology")
3. wykresdk(out, "EuroStandard")
}
